add env to skip git ignore and license

+++ zammad_new_szpm.rb
@@ -33,7 +33,7 @@ File.write(filename, %Q[{
 }])
 
 license_filename = "#{ARGV[0]}/LICENSE"
-if !File.exist?(license_filename) && (LICENSE == 'MIT' || LICENSE == 'GNU AFFERO GENERAL PUBLIC LICENSE')
+if !ENV['ZAMMAD_SCRIPTS_SKIP_LICENSE'] && !File.exist?(license_filename) && (LICENSE == 'MIT' || LICENSE == 'GNU AFFERO GENERAL PUBLIC LICENSE')
   if LICENSE == 'MIT'
     text = %Q[MIT License
 
@@ -726,7 +726,7 @@ For more information on this, and how to apply and follow the GNU AGPL, see
 end
 
 ignore_filename = "#{ARGV[0]}/.gitignore"
-if !File.exist?(ignore_filename)
+if !ENV['ZAMMAD_SCRIPTS_SKIP_GIT_IGNORE'] && !File.exist?(ignore_filename)
   text = Net::HTTP.get(URI.parse('https://raw.githubusercontent.com/zammad/zammad/develop/.gitignore'))
 
   File.write(ignore_filename, text)
